requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
kagglehub==0.3.4
mysql-connector-python==9.1.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
